Remove /ask and dead code cleanup
