+++ nucliadb/src/nucliadb/search/api/v1/__init__.py
@@ -18,7 +18,6 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 from . import (  # noqa: F401
-    ask,
     augment,
     catalog,
     feedback,
@@ -32,7 +31,6 @@ from . import (  # noqa: F401
     suggest,
     summarize,
 )
-from .resource import ask as ask_resource  # noqa: F401
 from .resource import ingestion_agents as ingestion_agents_resource  # noqa: F401
 from .resource import search as search_resource  # noqa: F401
 from .router import api  # noqa: F401
