--- nucliadb/src/nucliadb/search/search/chat/images.py
@@ -1,45 +1,0 @@
-# Copyright (C) 2021 Bosutech XXI S.L.
-#
-# nucliadb is offered under the AGPL v3.0 and as commercial software.
-# For commercial licensing, contact us at info@nuclia.com.
-#
-# AGPL:
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-import base64
-from io import BytesIO
-
-from nucliadb.ingest.fields.file import File
-from nucliadb.search import SERVICE_NAME
-from nucliadb_models.search import Image
-from nucliadb_utils.storages.storage import Storage
-from nucliadb_utils.utilities import get_storage
-
-
-# DEPRECATED(decoupled-ask): remove once old_prompt.py is removed
-async def get_file_thumbnail_image(file: File) -> Image | None:
-    fed = await file.get_file_extracted_data()
-    if fed is None or not fed.HasField("file_thumbnail"):
-        return None
-    storage: Storage = await get_storage(service_name=SERVICE_NAME)
-    image_bytes: BytesIO = await storage.downloadbytescf(fed.file_thumbnail)
-    value = image_bytes.getvalue()
-    if len(value) == 0:
-        return None
-    image = Image(
-        b64encoded=base64.b64encode(value).decode(),
-        # We assume the thumbnail is always generated as jpeg by Nuclia processing
-        content_type="image/jpeg",
-    )
-    return image
