--- nucliadb/src/nucliadb/search/search/chat/query.py
@@ -1,457 +1,0 @@
-# Copyright (C) 2021 Bosutech XXI S.L.
-#
-# nucliadb is offered under the AGPL v3.0 and as commercial software.
-# For commercial licensing, contact us at info@nuclia.com.
-#
-# AGPL:
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program. If not, see <http://www.gnu.org/licenses/>.
-#
-import asyncio
-from collections.abc import AsyncGenerator, Iterable
-
-from nidx_protos.nodereader_pb2 import GraphSearchResponse
-from nuclia_models.predict.generative_responses import GenerativeChunk
-
-from nucliadb.common.models_utils import to_proto
-from nucliadb.search import logger
-from nucliadb.search.predict import AnswerStatusCode, RephraseResponse
-from nucliadb.search.requesters.utils import Method, nidx_query
-from nucliadb.search.search.chat.exceptions import NoRetrievalResultsError
-from nucliadb.search.search.exceptions import IncompleteFindResultsError
-from nucliadb.search.search.find import find
-from nucliadb.search.search.merge import merge_relations_results
-from nucliadb.search.search.metrics import Metrics
-from nucliadb.search.search.query_parser.fetcher import Fetcher
-from nucliadb.search.search.query_parser.models import Query, RelationQuery, UnitRetrieval
-from nucliadb.search.search.query_parser.parsers.unit_retrieval import convert_retrieval_to_proto
-from nucliadb.search.search.rerankers import Reranker, get_reranker
-from nucliadb.search.settings import settings
-from nucliadb.search.utilities import get_predict
-from nucliadb_models import filters
-from nucliadb_models.search import (
-    AskRequest,
-    ChatContextMessage,
-    ChatModel,
-    ChatOptions,
-    FindOptions,
-    FindRequest,
-    KnowledgeboxFindResults,
-    NucliaDBClientType,
-    PreQueriesStrategy,
-    PreQuery,
-    PreQueryResult,
-    PromptContext,
-    PromptContextOrder,
-    Relations,
-    RephraseModel,
-    parse_rephrase_prompt,
-)
-from nucliadb_protos import audit_pb2
-from nucliadb_protos.utils_pb2 import RelationNode
-from nucliadb_telemetry.errors import capture_exception
-from nucliadb_utils.utilities import get_audit
-
-NOT_ENOUGH_CONTEXT_ANSWER = "Not enough data to answer this."
-
-
-async def rephrase_query(
-    kbid: str,
-    chat_history: list[ChatContextMessage],
-    query: str,
-    user_id: str,
-    user_context: list[str],
-    generative_model: str | None = None,
-    chat_history_relevance_threshold: float | None = None,
-) -> RephraseResponse:
-    # NOTE: When moving /ask to RAO, this will need to change to whatever client/utility is used
-    # to call NUA predict (internally or externally in the case of onprem).
-    predict = get_predict()
-    req = RephraseModel(
-        question=query,
-        chat_history=chat_history,
-        user_id=user_id,
-        user_context=user_context,
-        generative_model=generative_model,
-        chat_history_relevance_threshold=chat_history_relevance_threshold,
-    )
-    return await predict.rephrase_query(kbid, req)
-
-
-async def get_find_results(
-    *,
-    kbid: str,
-    query: str,
-    item: AskRequest,
-    ndb_client: NucliaDBClientType,
-    user: str,
-    origin: str,
-    metrics: Metrics,
-    prequeries_strategy: PreQueriesStrategy | None = None,
-) -> tuple[KnowledgeboxFindResults, list[PreQueryResult] | None, Fetcher, Reranker]:
-    prequeries_results = None
-    prefilter_queries_results = None
-    queries_results = None
-    if prequeries_strategy is not None:
-        prefilters = [prequery for prequery in prequeries_strategy.queries if prequery.prefilter]
-        prequeries = [prequery for prequery in prequeries_strategy.queries if not prequery.prefilter]
-        if len(prefilters) > 0:
-            with metrics.time("prefilters"):
-                prefilter_queries_results = await run_prequeries(
-                    kbid,
-                    prefilters,
-                    x_ndb_client=ndb_client,
-                    x_nucliadb_user=user,
-                    x_forwarded_for=origin,
-                    metrics=metrics.child_span("prefilters"),
-                )
-                prefilter_matching_resources = {
-                    resource
-                    for _, find_results in prefilter_queries_results
-                    for resource in find_results.resources.keys()
-                }
-                if len(prefilter_matching_resources) == 0:
-                    raise NoRetrievalResultsError()
-                # Make sure the main query and prequeries use the same resource filters.
-                # This is important to avoid returning results that don't match the prefilter.
-                matching_resources = list(prefilter_matching_resources)
-                add_resource_filter(item, matching_resources)
-                for prequery in prequeries:
-                    add_resource_filter(prequery.request, matching_resources)
-                    prequery.request.show_hidden = item.show_hidden
-
-        if prequeries:
-            with metrics.time("prequeries"):
-                queries_results = await run_prequeries(
-                    kbid,
-                    prequeries,
-                    x_ndb_client=ndb_client,
-                    x_nucliadb_user=user,
-                    x_forwarded_for=origin,
-                    metrics=metrics.child_span("prequeries"),
-                )
-
-        prequeries_results = (prefilter_queries_results or []) + (queries_results or [])
-
-    with metrics.time("main_query"):
-        main_results, fetcher, reranker = await run_main_query(
-            kbid,
-            query,
-            item,
-            ndb_client,
-            user,
-            origin,
-            metrics=metrics.child_span("main_query"),
-        )
-    return main_results, prequeries_results, fetcher, reranker
-
-
-def add_resource_filter(request: FindRequest | AskRequest, resources: list[str]):
-    if len(resources) == 0:
-        return
-
-    if request.filter_expression is not None:
-        if len(resources) > 1:
-            resource_filter: filters.FieldFilterExpression = filters.Or.model_validate(
-                {"or": [filters.Resource(prop="resource", id=rid) for rid in resources]}
-            )
-        else:
-            resource_filter = filters.Resource(prop="resource", id=resources[0])
-
-        # Add to filter expression if set
-        if request.filter_expression.field is None:
-            request.filter_expression.field = resource_filter
-        else:
-            request.filter_expression.field = filters.And.model_validate(
-                {"and": [request.filter_expression.field, resource_filter]}
-            )
-    else:
-        # Add to old key filters instead
-        request.resource_filters = resources
-
-
-def find_request_from_ask_request(item: AskRequest, query: str) -> FindRequest:
-    find_request = FindRequest()
-    find_request.filter_expression = item.filter_expression
-    find_request.resource_filters = item.resource_filters
-    find_request.features = []
-    if ChatOptions.SEMANTIC in item.features:
-        find_request.features.append(FindOptions.SEMANTIC)
-    if ChatOptions.KEYWORD in item.features:
-        find_request.features.append(FindOptions.KEYWORD)
-    if ChatOptions.RELATIONS in item.features:
-        find_request.features.append(FindOptions.RELATIONS)
-    find_request.query = query
-    find_request.fields = item.fields
-    find_request.filters = item.filters
-    find_request.field_type_filter = item.field_type_filter
-    find_request.min_score = item.min_score
-    find_request.vectorset = item.vectorset
-    find_request.range_creation_start = item.range_creation_start
-    find_request.range_creation_end = item.range_creation_end
-    find_request.range_modification_start = item.range_modification_start
-    find_request.range_modification_end = item.range_modification_end
-    find_request.show = item.show
-    find_request.extracted = item.extracted
-    find_request.highlight = item.highlight
-    find_request.security = item.security
-    find_request.debug = item.debug
-    find_request.rephrase = item.rephrase
-    find_request.rephrase_prompt = parse_rephrase_prompt(item)
-    find_request.query_image = item.query_image
-    find_request.rank_fusion = item.rank_fusion
-    find_request.reranker = item.reranker
-    # We don't support pagination, we always get the top_k results.
-    find_request.top_k = item.top_k
-    find_request.show_hidden = item.show_hidden
-    find_request.generative_model = item.generative_model
-
-    # this executes the model validators, that can tweak some fields
-    return FindRequest.model_validate(find_request)
-
-
-async def run_main_query(
-    kbid: str,
-    query: str,
-    item: AskRequest,
-    ndb_client: NucliaDBClientType,
-    user: str,
-    origin: str,
-    metrics: Metrics,
-) -> tuple[KnowledgeboxFindResults, Fetcher, Reranker]:
-    find_request = find_request_from_ask_request(item, query)
-
-    find_results, incomplete, fetcher, reranker = await find_retrieval(
-        kbid,
-        find_request,
-        ndb_client,
-        user,
-        origin,
-        metrics=metrics,
-    )
-    if incomplete:
-        raise IncompleteFindResultsError()
-    return find_results, fetcher, reranker
-
-
-async def get_relations_results(
-    *,
-    kbid: str,
-    text_answer: str,
-    timeout: float | None = None,
-) -> Relations:
-    try:
-        predict = get_predict()
-        detected_entities = await predict.detect_entities(kbid, text_answer)
-
-        return await get_relations_results_from_entities(
-            kbid=kbid,
-            entities=detected_entities,
-            timeout=timeout,
-        )
-    except Exception as exc:
-        capture_exception(exc)
-        logger.exception("Error getting relations results")
-        return Relations(entities={})
-
-
-async def get_relations_results_from_entities(
-    *,
-    kbid: str,
-    entities: Iterable[RelationNode],
-    timeout: float | None = None,
-    deleted_entities: set[str] = set(),
-) -> Relations:
-    entry_points = list(entities)
-    retrieval = UnitRetrieval(
-        query=Query(
-            relation=RelationQuery(
-                entry_points=entry_points,
-                deleted_entities={"": list(deleted_entities)},
-                deleted_entity_groups=[],
-            )
-        ),
-        top_k=50,
-    )
-    request = convert_retrieval_to_proto(retrieval)
-
-    result = await nidx_query(
-        kbid,
-        Method.SEARCH,
-        request,
-        timeout=timeout,
-    )
-    relations_results: GraphSearchResponse = result.graph
-    return await merge_relations_results(
-        relations_results,
-        entry_points,
-    )
-
-
-def maybe_audit_chat(
-    *,
-    kbid: str,
-    user_id: str,
-    client_type: NucliaDBClientType,
-    origin: str,
-    generative_answer_time: float,
-    generative_answer_first_chunk_time: float,
-    generative_reasoning_first_chunk_time: float | None,
-    rephrase_time: float | None,
-    user_query: str,
-    rephrased_query: str | None,
-    retrieval_rephrase_query: str | None,
-    text_answer: bytes,
-    text_reasoning: str | None,
-    status_code: AnswerStatusCode,
-    chat_history: list[ChatContextMessage],
-    query_context: PromptContext,
-    query_context_order: PromptContextOrder,
-    learning_id: str | None,
-    model: str | None,
-):
-    audit = get_audit()
-    if audit is None:
-        return
-
-    audit_answer = parse_audit_answer(text_answer, status_code)
-    # Append chat history
-    chat_history_context = [
-        audit_pb2.ChatContext(author=message.author, text=message.text) for message in chat_history
-    ]
-
-    # Append paragraphs retrieved on this chat
-    chat_retrieved_context = [
-        audit_pb2.RetrievedContext(text_block_id=paragraph_id, text=text)
-        for paragraph_id, text in query_context.items()
-    ]
-
-    audit.chat(
-        kbid,
-        user_id,
-        to_proto.client_type(client_type),
-        origin,
-        question=user_query,
-        generative_answer_time=generative_answer_time,
-        generative_answer_first_chunk_time=generative_answer_first_chunk_time,
-        generative_reasoning_first_chunk_time=generative_reasoning_first_chunk_time,
-        rephrase_time=rephrase_time,
-        rephrased_question=rephrased_query,
-        retrieval_rephrased_question=retrieval_rephrase_query,
-        chat_context=chat_history_context,
-        retrieved_context=chat_retrieved_context,
-        answer=audit_answer,
-        reasoning=text_reasoning,
-        learning_id=learning_id,
-        status_code=int(status_code.value),
-        model=model,
-    )
-
-
-def parse_audit_answer(raw_text_answer: bytes, status_code: AnswerStatusCode) -> str | None:
-    if status_code == AnswerStatusCode.NO_CONTEXT or status_code == AnswerStatusCode.NO_RETRIEVAL_DATA:
-        # We don't want to audit "Not enough context to answer this." and instead set a None.
-        return None
-    return raw_text_answer.decode()
-
-
-def tokens_to_chars(n_tokens: int) -> int:
-    # Multiply by 3 to have a good margin and guess between characters and tokens.
-    # This will be properly cut at the NUA predict API.
-    return n_tokens * 3
-
-
-def sorted_prompt_context_list(context: PromptContext, order: PromptContextOrder) -> list[str]:
-    """
-    context = {"x": "foo", "y": "bar"}
-    order = {"y": 1, "x": 0}
-    sorted_prompt_context_list(context, order) == ["foo", "bar"]
-    """
-    sorted_items = sorted(
-        context.items(),
-        key=lambda item: order.get(item[0], float("inf")),
-    )
-    return list(map(lambda item: item[1], sorted_items))
-
-
-async def run_prequeries(
-    kbid: str,
-    prequeries: list[PreQuery],
-    x_ndb_client: NucliaDBClientType,
-    x_nucliadb_user: str,
-    x_forwarded_for: str,
-    metrics: Metrics,
-) -> list[PreQueryResult]:
-    """
-    Runs simultaneous find requests for each prequery and returns the merged results according to the normalized weights.
-    """
-    results: list[PreQueryResult] = []
-    max_parallel_prequeries = asyncio.Semaphore(settings.prequeries_max_parallel)
-
-    async def _prequery_find(prequery: PreQuery, index: int):
-        async with max_parallel_prequeries:
-            prequery_id = prequery.id or f"prequery-{index}"
-            find_results, _, _, _ = await find_retrieval(
-                kbid,
-                prequery.request,
-                x_ndb_client,
-                x_nucliadb_user,
-                x_forwarded_for,
-                metrics=metrics.child_span(prequery_id),
-            )
-            return prequery, find_results
-
-    ops = []
-    for idx, prequery in enumerate(prequeries):
-        ops.append(asyncio.create_task(_prequery_find(prequery, idx)))
-    ops_results = await asyncio.gather(*ops)
-    for prequery, find_results in ops_results:
-        results.append((prequery, find_results))
-    return results
-
-
-async def get_answer_stream(
-    kbid: str,
-    item: ChatModel,
-    extra_headers: dict[str, str] | None = None,
-) -> tuple[str, str, AsyncGenerator[GenerativeChunk, None]]:
-    # NOTE: When moving /ask to RAO, this will need to change to whatever client/utility is used
-    # to call NUA predict (internally or externally in the case of onprem).
-    predict = get_predict()
-    return await predict.chat_query_ndjson(
-        kbid=kbid,
-        item=item,
-        extra_headers=extra_headers,
-    )
-
-
-async def find_retrieval(
-    kbid: str,
-    find_request: FindRequest,
-    x_ndb_client: NucliaDBClientType,
-    x_nucliadb_user: str,
-    x_forwarded_for: str,
-    metrics: Metrics,
-) -> tuple[KnowledgeboxFindResults, bool, Fetcher, Reranker]:
-    results, incomplete, parsed = await find(
-        kbid,
-        find_request,
-        x_ndb_client,
-        x_nucliadb_user,
-        x_forwarded_for,
-        metrics=metrics,
-    )
-    # this has already been asserted inside the find() call
-    assert parsed.retrieval.reranker is not None, "find parser must provide a reranking algorithm"
-    reranker = get_reranker(parsed.retrieval.reranker)
-    return results, incomplete, parsed.fetcher, reranker
