+++ nucliadb/src/nucliadb/search/search/query_parser/fetcher.py
@@ -32,7 +32,7 @@ from nucliadb.search.predict_models import QueryModel
 from nucliadb.search.search.metrics import query_parse_dependency_observer
 from nucliadb.search.utilities import get_predict
 from nucliadb_models.internal.predict import QueryInfo
-from nucliadb_models.search import Image, MaxTokens
+from nucliadb_models.search import Image
 from nucliadb_protos import knowledgebox_pb2, utils_pb2
 
 
@@ -291,35 +291,6 @@ class Fetcher:
             synonyms = await get_kb_synonyms(self.kbid)
             self.cache.synonyms = synonyms
             return synonyms
-
-    # Generative
-
-    async def get_visual_llm_enabled(self) -> bool:
-        query_info = await self._predict_query_endpoint()
-        if query_info is None:
-            raise SendToPredictError("Error while using predict's query endpoint")
-
-        return query_info.visual_llm
-
-    async def get_max_context_tokens(self, max_tokens: MaxTokens | None) -> int:
-        query_info = await self._predict_query_endpoint()
-        if query_info is None:
-            raise SendToPredictError("Error while using predict's query endpoint")
-
-        model_max = query_info.max_context
-        if max_tokens is not None and max_tokens.context is not None:
-            if max_tokens.context > model_max:
-                raise InvalidQueryError(
-                    "max_tokens.context",
-                    f"Max context tokens is higher than the model's limit of {model_max}",
-                )
-            return max_tokens.context
-        return model_max
-
-    def get_max_answer_tokens(self, max_tokens: MaxTokens | None) -> int | None:
-        if max_tokens is not None and max_tokens.answer is not None:
-            return max_tokens.answer
-        return None
 
     # Predict API
