+++ nucliadb/tests/search/integration/search/test_ask_compat_image_strategies.py
